+++ src/raps/actions/swap.ts
@@ -1,6 +1,10 @@
 import { Wallet } from '@ethersproject/wallet';
 import { captureException } from '@sentry/react-native';
-import { Rap, RapActionParameters, SwapActionParameters } from '../common';
+import {
+  Rap,
+  RapEchangeActionParameters,
+  SwapActionParameters,
+} from '../common';
 import {
   ProtocolType,
   TransactionStatus,
@@ -24,7 +28,7 @@ const swap = async (
   wallet: Wallet,
   currentRap: Rap,
   index: number,
-  parameters: RapActionParameters,
+  parameters: RapEchangeActionParameters,
   baseNonce?: number
 ): Promise<number | undefined> => {
   logger.log(`[${actionName}] base nonce`, baseNonce, 'index:', index);
