+++ src/raps/actions/unlock.ts
@@ -4,7 +4,11 @@ import { Wallet } from '@ethersproject/wallet';
 import { captureException } from '@sentry/react-native';
 import { isNull, toLower } from 'lodash';
 import { alwaysRequireApprove } from '../../config/debug';
-import { Rap, RapActionParameters, UnlockActionParameters } from '../common';
+import {
+  Rap,
+  RapEchangeActionParameters,
+  UnlockActionParameters,
+} from '../common';
 import {
   Asset,
   TransactionStatus,
@@ -82,7 +86,7 @@ const unlock = async (
   wallet: Wallet,
   currentRap: Rap,
   index: number,
-  parameters: RapActionParameters,
+  parameters: RapEchangeActionParameters,
   baseNonce?: number
 ): Promise<number | undefined> => {
   logger.log(`[${actionName}] base nonce`, baseNonce, 'index:', index);
