+++ src/raps/actions/withdrawCompound.ts
@@ -1,7 +1,11 @@
 import { Contract } from '@ethersproject/contracts';
 import { Wallet } from '@ethersproject/wallet';
 import { captureException } from '@sentry/react-native';
-import { Rap, RapActionParameters, SwapActionParameters } from '../common';
+import {
+  Rap,
+  RapEchangeActionParameters,
+  SwapActionParameters,
+} from '../common';
 import {
   ProtocolType,
   TransactionStatus,
@@ -30,7 +34,7 @@ const withdrawCompound = async (
   wallet: Wallet,
   currentRap: Rap,
   index: number,
-  parameters: RapActionParameters,
+  parameters: RapEchangeActionParameters,
   baseNonce?: number
 ): Promise<number | undefined> => {
   logger.log(`[${actionName}] base nonce`, baseNonce, 'index:', index);
